+++ kubernetes-tests/src/test/java/io/fabric8/openshift/client/server/mock/NamespacedItemTest.java
@@ -22,6 +22,7 @@ import io.fabric8.kubernetes.api.model.ConfigMapList;
 import io.fabric8.kubernetes.api.model.HasMetadata;
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.fabric8.kubernetes.client.NamespacedKubernetesClient;
+import io.fabric8.kubernetes.client.dsl.ConfigMapResource;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.NamespaceListVisitFromServerGetDeleteRecreateWaitApplicable;
 import io.fabric8.kubernetes.client.dsl.Resource;
@@ -118,7 +119,7 @@ class NamespacedItemTest {
 
   @Test
   void testOperationNullNamespace() {
-    MixedOperation<ConfigMap, ConfigMapList, Resource<ConfigMap>> configMaps = this.client.configMaps();
+    MixedOperation<ConfigMap, ConfigMapList, ConfigMapResource> configMaps = this.client.configMaps();
     assertThrows(KubernetesClientException.class, () -> configMaps.inNamespace(null));
   }